Render the planets
This is the first time I’m rendering something
from the Model in the View.

+++ src/model/planet/mod.rs
@@ -1,5 +1,4 @@
-use super::point::Point;
-use super::circle::Circle;
+use ::util::{Circle, Point};
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct Planet {
@@ -9,14 +8,14 @@ pub struct Planet {
 }
 
 impl Planet {
-    pub fn new(location: Point, radius: f64, zone_height: f64, ordinal: usize) -> Self {
+    pub fn new(location: Point, radius: f32, zone_height: f32, ordinal: usize) -> Self {
         let mass = Circle::new(location, radius);
         let zone = Circle::new(location, Self::zone_height(radius, zone_height));
 
         Planet { mass, zone, ordinal }
     }
 
-    pub fn zone_height(radius: f64, height: f64) -> f64 {
+    pub fn zone_height(radius: f32, height: f32) -> f32 {
         radius * (height + 1.0)
     }
 }
